pyproject.toml
[tool.poetry]
name = "voiceai-v2"
version = "0.1.0"
description = "speech-to-text"
authors = ["genagurbanguliyev"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.109.0"
uvicorn = {extras = ["standart"], version = "^0.27.0.post1"}
pydantic = "^2.6.0"
pydantic-settings = "^2.1.0"
python-dotenv = "^1.0.1"
sqlalchemy = "^2.0.25"
psycopg = {extras = ["binary"], version = "^3.1.17"}
alembic = "^1.13.1"
python-jose = "^3.3.0"
python-multipart = "^0.0.7"
passlib = {extras = ["bcrypt"], version = "^1.7.4"}
hypercorn = {extras = ["trio"], version = "^0.16.0"}


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
